Veines multi-question steps now require an answer for every question.
Previously, users could select a single item and leave the others blank before being able to click Continue or Done. This would crash the app if done and no data would be sent to Firebase. Users now have to click an answer for every question in every step before continuing.